StockAI - Autonomous AI Trading Agent/requirements.txt
# StockAI - Advanced Multi-Agent Trading System
# Dependencies for the trading system

# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning & Reinforcement Learning
stable-baselines3>=2.0.0
gym>=0.21.0
joblib>=1.2.0

# Technical Analysis
TA-Lib>=0.4.25

# API & Web Communication
alpaca-trade-api>=3.0.0
websocket-client>=1.4.0
requests>=2.28.0

# Data Processing & Analysis
scikit-learn>=1.1.0
scipy>=1.9.0

# Natural Language Processing (for sentiment analysis)
textblob>=0.17.0

# Configuration & Utilities
pydantic>=1.10.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.8.0

# Caching & Message Queue
redis>=4.3.0
celery>=5.2.0

# Web Framework (for future API endpoints)
fastapi>=0.85.0
uvicorn>=0.18.0

# Streamlit UI
streamlit>=1.28.0
plotly>=5.15.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
isort>=5.10.0

# Monitoring & Logging
prometheus-client>=0.14.0
structlog>=22.1.0

# Optional: Jupyter for development
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
